refactor(#1704): campsite split — move sequential.rs inline tests to a sibling file
Verbatim move (#1116/#1135). sequential.rs was 1319 lines; it is now 827.
